nit: relocate helpers
